Make gdb support test lazy
