+++ numba/tests/gdb_support.py
@@ -183,29 +183,3 @@ class GdbMIDriver(object):
         g = m.groups()
         assert len(g) == 1, "Invalid number of match groups found"
         return g[0].replace('"', '').split(',')
-
-
-def _gdb_has_python():
-    if _HAVE_PEXPECT and _HAVE_GDB:
-        driver = GdbMIDriver(__file__, debug=False,)
-        has_python = driver.supports_python()
-        driver.quit()
-        return has_python
-    else:
-        return False
-
-
-def _gdb_has_numpy():
-    if _HAVE_PEXPECT and _HAVE_GDB:
-        driver = GdbMIDriver(__file__, debug=False,)
-        has_numpy = driver.supports_numpy()
-        driver.quit()
-        return has_numpy
-    else:
-        return False
-
-
-_msg = "functioning gdb present, but it does not python"
-skip_unless_gdb_has_python = unittest.skipUnless(_gdb_has_python(), _msg)
-_msg = "functioning gdb present and has Python, but it does not have NumPy"
-skip_unless_gdb_has_numpy = unittest.skipUnless(_gdb_has_numpy(), _msg)
